+++ src/modules/Image/encodings/x.c
@@ -271,12 +271,12 @@ static void image_x_encode_truecolor(INT32 args)
    }
    else
    {
-      INT32 rfshift,gfshift,bfshift,rzshift,gzshift,bzshift;
-      INT32 bpshift,blinemod,bit;
+      ptrdiff_t rfshift,gfshift,bfshift,rzshift,gzshift,bzshift;
+      ptrdiff_t bpshift,blinemod,bit;
 
-      rfshift=rshift-(sizeof(COLORTYPE)*8-rbits);
-      gfshift=gshift-(sizeof(COLORTYPE)*8-gbits);
-      bfshift=bshift-(sizeof(COLORTYPE)*8-bbits);
+      rfshift = rshift-(sizeof(COLORTYPE)*8-rbits);
+      gfshift = gshift-(sizeof(COLORTYPE)*8-gbits);
+      bfshift = bshift-(sizeof(COLORTYPE)*8-bbits);
       if (rfshift<0) rzshift=-rfshift,rfshift=0; else rzshift=0;
       if (gfshift<0) gzshift=-gfshift,gfshift=0; else gzshift=0;
       if (bfshift<0) bzshift=-bfshift,bfshift=0; else bzshift=0;
